+++ perf/FormattedFiles.cs
@@ -1,6 +1,10 @@
-﻿using System.Collections.Immutable;
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the MIT license.  See License.txt in the project root for license information.
+
+using System;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Jobs;
+using Microsoft.CodeAnalysis.Tools.Utilities;
+using Microsoft.Extensions.FileSystemGlobbing;
 using Microsoft.Extensions.Logging;
 
 namespace Microsoft.CodeAnalysis.Tools.Perf
@@ -8,10 +12,12 @@ namespace Microsoft.CodeAnalysis.Tools.Perf
     [SimpleJob(RuntimeMoniker.NetCoreApp21)]
     public class FormattedFiles
     {
-        private const string UnformattedProjectPath = "tests/projects/for_code_formatter/unformatted_project";
-        private const string UnformattedProjectFilePath = UnformattedProjectPath + "/unformatted_project.csproj";
+        private const string UnformattedProjectPath = "tests/projects/for_code_formatter/unformatted_project/";
+        private const string UnformattedProjectFilePath = UnformattedProjectPath + "unformatted_project.csproj";
         private const string UnformattedSolutionFilePath = "tests/projects/for_code_formatter/unformatted_solution/unformatted_solution.sln";
-        private static EmptyLogger EmptyLogger = new EmptyLogger();
+
+        private static readonly EmptyLogger EmptyLogger = new EmptyLogger();
+        private static readonly Matcher AllFileMatcher = SourceFileMatcher.CreateMatcher(Array.Empty<string>(), Array.Empty<string>());
 
         [IterationSetup]
         public void NoFilesFormattedSetup()
@@ -30,8 +36,7 @@ namespace Microsoft.CodeAnalysis.Tools.Perf
                 LogLevel.Error,
                 saveFormattedFiles: false,
                 changesAreErrors: false,
-                ImmutableHashSet<string>.Empty,
-                ImmutableHashSet<string>.Empty,
+                AllFileMatcher,
                 reportPath: string.Empty);
             _ = CodeFormatter.FormatWorkspaceAsync(options, EmptyLogger, default).GetAwaiter().GetResult();
         }
@@ -46,8 +51,7 @@ namespace Microsoft.CodeAnalysis.Tools.Perf
                 LogLevel.Error,
                 saveFormattedFiles: false,
                 changesAreErrors: false,
-                ImmutableHashSet<string>.Empty,
-                ImmutableHashSet<string>.Empty,
+                AllFileMatcher,
                 reportPath: string.Empty);
             _ = CodeFormatter.FormatWorkspaceAsync(options, EmptyLogger, default).GetAwaiter().GetResult();
         }
@@ -62,8 +66,7 @@ namespace Microsoft.CodeAnalysis.Tools.Perf
                 LogLevel.Error,
                 saveFormattedFiles: false,
                 changesAreErrors: false,
-                ImmutableHashSet<string>.Empty,
-                ImmutableHashSet<string>.Empty,
+                AllFileMatcher,
                 reportPath: string.Empty);
             _ = CodeFormatter.FormatWorkspaceAsync(options, EmptyLogger, default).GetAwaiter().GetResult();
         }