+++ perf/Utilities/EmptyLogger.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the MIT license.  See License.txt in the project root for license information.
+
+using System;
+using Microsoft.CodeAnalysis.Tools.Logging;
 using Microsoft.Extensions.Logging;
 
 namespace Microsoft.CodeAnalysis.Tools.Perf
@@ -16,13 +19,6 @@ namespace Microsoft.CodeAnalysis.Tools.Perf
 
         public bool IsEnabled(LogLevel logLevel) => false;
 
-        public IDisposable BeginScope<TState>(TState state) => new EmptyScope();
-
-        private class EmptyScope : IDisposable
-        {
-            public void Dispose()
-            {
-            }
-        }
+        public IDisposable BeginScope<TState>(TState state) => NullScope.Instance;
     }
 }