+++ perf/Utilities/MSBuildRegister.cs
@@ -1,4 +1,6 @@
-﻿using System.Linq;
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the MIT license.  See License.txt in the project root for license information.
+
+using System.Linq;
 using System.Threading;
 using Microsoft.CodeAnalysis.Tools.MSBuild;