+++ perf/Utilities/WorkspacePathHelper.cs
@@ -1,4 +1,6 @@
-﻿using System.IO;
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the MIT license.  See License.txt in the project root for license information.
+
+using System.IO;
 
 namespace Microsoft.CodeAnalysis.Tools.Perf
 {