+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All changes to the project will be documented in this file.
 ## [4.0.x] - Not Yet Released
 ### Merged PRs:
 - [Combine --check and --dry-run into a single option. (541)](https://github.com/dotnet/format/pull/541)
+- [Use space-separated paths instead of comma-separated for --include and --exclude (551)](https://github.com/dotnet/format/pull/551)
+- [Support loading commandline options from response files (552)](https://github.com/dotnet/format/pull/552)
+- [Support file globbing in --include and --exclude options (555)](https://github.com/dotnet/format/pull/555)
 
 ## [3.3.111304] - 2020-02-13
 [View Complete Diff of Changes](https://github.com/dotnet/format/compare/3ecea99de4bb82b724bf11134279b5aaf8dd1f2f...7c8f67a570f5fde6a247704733d6742f93c0fa48)